move event bus to own file
