+++ timesketch/frontend-ng/src/App.vue
@@ -29,7 +29,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from './main'
+import EventBus from './event-bus.js'
 
 export default {
   name: 'app',
