+++ timesketch/frontend-ng/src/components/Explore/Comments.vue
@@ -86,7 +86,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 import ApiClient from '../../utils/RestApiClient'
 
 export default {
