+++ timesketch/frontend-ng/src/components/Explore/EventActionMenu.vue
@@ -45,7 +45,7 @@ limitations under the License.
 
 <script>
 import ApiClient from '../../utils/RestApiClient'
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   props: ['event'],
