+++ timesketch/frontend-ng/src/components/Explore/EventDetail.vue
@@ -169,7 +169,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 import ApiClient from '../../utils/RestApiClient'
 import TsAggregateDialog from './AggregateDialog.vue'
 import TsFormatXmlString from './FormatXMLString.vue'
