+++ timesketch/frontend-ng/src/components/Explore/EventList.vue
@@ -418,7 +418,7 @@ limitations under the License.
 
 <script>
 import ApiClient from '../../utils/RestApiClient.js'
-import EventBus from '../../main.js'
+import EventBus from '../../event-bus.js'
 
 import TsBarChart from './BarChart.vue'
 import TsEventDetail from './EventDetail.vue'
