+++ timesketch/frontend-ng/src/components/Explore/SearchHistoryButtons.vue
@@ -30,7 +30,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   computed: {
