+++ timesketch/frontend-ng/src/components/Graph/Cytoscape.vue
@@ -165,7 +165,7 @@ limitations under the License.
 
 <script>
 import ApiClient from '../../utils/RestApiClient.js'
-import EventBus from '../../main.js'
+import EventBus from '../../event-bus.js'
 import cytoscape from 'cytoscape'
 import spread from 'cytoscape-spread'
 import dagre from 'cytoscape-dagre'
