+++ timesketch/frontend-ng/src/components/LeftPanel/DataTypes.vue
@@ -75,7 +75,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   props: [],
