+++ timesketch/frontend-ng/src/components/LeftPanel/Graphs.vue
@@ -78,7 +78,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   props: [],
