+++ timesketch/frontend-ng/src/components/LeftPanel/SavedSearches.vue
@@ -82,7 +82,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   props: [],
