+++ timesketch/frontend-ng/src/components/LeftPanel/TagsList.vue
@@ -64,7 +64,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 export default {
   props: [],
