+++ timesketch/frontend-ng/src/components/Scenarios/SearchChip.vue
@@ -30,7 +30,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../../main'
+import EventBus from '../../event-bus.js'
 
 const defaultQueryFilter = () => {
   return {
