+++ timesketch/frontend-ng/src/views/Explore.vue
@@ -306,7 +306,7 @@ limitations under the License.
 
 <script>
 import ApiClient from '../utils/RestApiClient.js'
-import EventBus from '../main.js'
+import EventBus from '../event-bus.js'
 
 import { dragscroll } from 'vue-dragscroll'
