+++ timesketch/frontend-ng/src/views/Graph.vue
@@ -28,7 +28,7 @@ limitations under the License.
 </template>
 
 <script>
-import EventBus from '../main.js'
+import EventBus from '../event-bus.js'
 import TsCytoscape from '../components/Graph/Cytoscape.vue'
 
 export default {
