+++ timesketch/frontend-ng/src/views/ThreatIntel.vue
@@ -86,7 +86,7 @@ limitations under the License.
 
 <script>
 import ApiClient from '../utils/RestApiClient.js'
-import EventBus from '../main.js'
+import EventBus from '../event-bus.js'
 import TsIndicatorDialog from '../components/ThreatIntel/IndicatorDialog.vue'
 
 const defaultQueryFilter = () => {
